- There were lots of conflicts in the merge with the last 70 commits
to HEAD.
- There were also lots of warnings, some in the Defensive
code, and some in Dyninst, that I fixed.
- The build issues were due to Windows-isms creeping into the
codebase.
- The testsuite doesn't build, but Drew was looking into that in
a separate commit

Merge branch 'master' of ssh://git.dyninst.org/pub/dyninst into Kevin-Drew

Conflicts:
	dataflowAPI/h/SymEval.h
	dataflowAPI/src/SymEval.C
	dataflowAPI/src/slicing.C
	dataflowAPI/src/stackanalysis.C
	dyninstAPI/h/BPatch_addressSpace.h
	dyninstAPI/h/BPatch_function.h
	dyninstAPI/h/BPatch_snippet.h
	dyninstAPI/src/BPatch_addressSpace.C
	dyninstAPI/src/ast.C
	dyninstAPI/src/function.C
	parseAPI/src/IA_IAPI.h
	parseAPI/src/ParseData.C
	parseAPI/src/ParseData.h
	parseAPI/src/Parser.C
	parseAPI/src/ParserDetails.C

+++ common/src/arch-sparc.C
@@ -758,7 +758,7 @@ void instruction::get_register_operands()
         }
 
   if (reads.size() > 8)
-     fprintf(stderr, "%s[%d]:  WARNING:  unexpected size %ld for read registers\n", FILE__, __LINE__, reads.size());
+     fprintf(stderr, "%s[%d]:  WARNING:  unexpected size %ld for read registers\n", FILE__, __LINE__, (long int) reads.size());
 
   for (unsigned int i = 0; i < reads.size(); ++i) {
      if (reads[i].getNumber() == -1)
@@ -772,7 +772,7 @@ void instruction::get_register_operands()
 
 
   if (writes.size() > 6)
-     fprintf(stderr, "%s[%d]:  WARNING:  unexpected size %ld for write registers\n", FILE__, __LINE__, writes.size());
+     fprintf(stderr, "%s[%d]:  WARNING:  unexpected size %ld for write registers\n", FILE__, __LINE__, (long int) writes.size());
 
   for (unsigned int i = 0; i < writes.size(); ++i) {
      if (writes[i].getNumber() == -1)
